requirements.txt
build>=1.0.0
PyYAML>=6.0.0